Source-GOP-aware constant-duration segments; compact DASH manifest
Probe each source's GOP cadence (first 60s of packets) and adopt the
source frame grid for the whole job when GOP is constant and ≤ 2s.
Every rendition then cuts at the same wall-clock instants: remuxed
top-resolution profiles inherit source IDRs, transcoded lower
profiles force-key-frame to the same time grid (per-profile re-snap
when fps_limit drops the rendition below source fps so each lands
honestly on its own output frame grid). Variable-GOP or > 2s sources
fall back to a fixed 2s GOP and 6s segments with no remux.

DASH manifest now picks per-Representation between
SegmentTemplate-with-duration (compact, used when each rendition's
segments are uniform within ±100ms of its modal value) and
SegmentTimeline (per-segment durations, used when the AAC encoder's
periodic drift makes audio non-uniform). Video and audio decide
independently, so most jobs collapse the video AdaptationSet to one
duration attribute per rendition while keeping audio honest. Timescale
is microseconds so the declared values stay tight on multi-hour
content.

Player-side: WatchPage.jsx now destroys the Shaka instance only on
404s for .m3u8 / .mpd / .mp4 (manifest + init segments); .m4s media
segment 404s defer to Shaka's built-in retry so a permanently deleted
mid-stream segment stalls at that point rather than killing the
session.

+++ worker/internal/config/config.go
@@ -21,8 +21,13 @@ import (
 // are passed through unchanged (no upsampling).
 //
 // GOPSeconds is the keyframe interval in seconds — the FFmpeg keyint flag is
-// computed as round(GOPSeconds * effective_fps) at arg-build time so GOP
-// stays a constant 2s regardless of source frame rate.
+// computed as ceil(GOPSeconds * effective_fps) at arg-build time and
+// -force_key_frames pins the cadence to multiples of GOPSeconds so every
+// rendition cuts at the same instants regardless of source frame rate.
+//
+// SegmentDuration is in seconds and accepts fractional values. The worker may
+// override this at runtime to align with a probed source GOP — see
+// slot/job.go runTranscode for the per-job override that propagates here.
 type OutputProfile struct {
 	Name             string  `json:"name"`
 	Width            int     `json:"width"`
@@ -32,7 +37,7 @@ type OutputProfile struct {
 	Codec            string  `json:"codec"`
 	Profile          string  `json:"profile"`
 	Preset           string  `json:"preset"`
-	SegmentDuration  int     `json:"segment_duration"`
+	SegmentDuration  float64 `json:"segment_duration"`
 	GOPSeconds       float64 `json:"gop_seconds"`
 }
